feat: Update index route to include new sections and remove under construction content
